requirements_m3fc.txt
aiohttp==3.8.5
aiohttp-cors==0.7.0
aioredis==1.3.1
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
blessed==1.20.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.0.0
colorama==0.4.6
colorful==0.5.5
contourpy==1.1.0
cycler==0.11.0
dm-tree==0.1.8
executing==1.2.0
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.0
gillespy2==1.5.7
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.60.0
gpustat==1.1.1
grpcio==1.57.0
# gym==0.20.0
gym-notices==0.0.8
hiredis==2.2.3
icecream==2.1.3
idna==3.4
imageio==2.31.1
importlib-metadata==4.13.0
importlib-resources==6.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lz4==4.3.2
# ma-gym==0.0.14
matplotlib==3.7.2
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numpy==1.20.3
nvidia-ml-py==12.535.77
opencensus==0.11.2
opencensus-context==0.1.3
opencv-python==3.4.18.65
packaging==23.1
pandas==2.0.3
PettingZoo==1.12.0
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
plotly==4.14.3
POT==0.8.2
prometheus-client==0.17.1
# protobuf==3.20.0
psutil==5.9.5
py-spy==0.3.14
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyglet==1.5.27
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
ray==1.8.0
redis==5.0.0
referencing==0.30.2
requests==2.31.0
retrying==1.3.3
rpds-py==0.9.2
rsa==4.9
scikit-image==0.19.3
scipy==1.10.1
six==1.16.0
SuperSuit==3.2.0
tabulate==0.9.0
tensorboardX==2.6.2.2
tifffile==2023.7.10
torch==1.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
yarl==1.9.2
zipp==3.16.2
